Change autocompleter API.
It is now possible to import `porcupine.plugins.autocompleter` and add
your own completers.

The completers take as arguments the contents of the current file, the
line the cursor is on and the column the cursor is on.
They must return an iterable that consists of the possible completions
for the current situation.

Possible completions means "rest of the word", "not full word". For
example, if you wanted to complete "string.ascii_upperc" to
"string.ascii_uppercase" you'd return `["ase"]`.

The `porcupine.plugins.autocompleter.completer` decorator allows you to
add completers by file-type, and it is called like this:

@completer("Python")
def jedi_complete(source, line, col):
    script = jedi.Script(source, line, col)
    return (c.complete for c in script.completions())